More emitting primitive information
