+++ include/rendering/IRenderable.hpp
@@ -17,13 +17,15 @@ struct EmittingInstanceData {
     glm::mat4 model_matrix;
     glm::vec4 emission;
     uint32_t instance_id;
-    uint32_t padding[3];
+    uint32_t primitive_count;
+    uint32_t padding[2];
 };
 
 struct RenderObject {
     InstanceData instance_data;
     std::shared_ptr<AccelerationStructure> acceleration_structure;
     glm::mat4 transform;
+    uint32_t primitive_count;
 };
 
 struct DrawContext {
