+++ src/rendering/nodes/MeshNode.cpp
@@ -7,7 +7,7 @@
 void MeshNode::draw(const glm::mat4 &topMatrix, DrawContext &ctx) {
     glm::mat4 nodeMatrix = topMatrix * worldTransform;
 
-    ctx.objects.push_back(RenderObject{InstanceData{meshAsset->geometry_id, meshMaterial->material_index}, meshAsset->accelerationStructure, nodeMatrix});
+    ctx.objects.push_back(RenderObject{InstanceData{meshAsset->geometry_id, meshMaterial->material_index}, meshAsset->accelerationStructure, nodeMatrix, meshAsset->triangle_count});
 
     Node::draw(topMatrix, ctx);
 }
